Move metadata path before dynamic app key
