+++ spec/apibuilder-api.json
@@ -1017,6 +1017,31 @@
       "operations": [
         {
           "method": "GET",
+          "description": "Returns the versions assocoated with the specified application. The latest version is the first result returned.",
+	  "path": "/metadata/:applicationKey/versions",
+          "parameters": [
+            { "name": "limit", "type": "long", "default": 25, "description": "The number of records to return", "maximum": 100 },
+            { "name": "offset", "type": "long", "default": 0, "description": "Used to paginate. First page of results is 0.", "minimum": 0 }
+          ],
+          "responses": {
+            "200": { "type": "[application_metadata_version]" },
+            "401": { "type": "unit" },
+            "404": { "type": "unit" }
+          }
+        },
+        {
+          "method": "GET",
+          "description": "Returns the latest version number as a string",
+	  "path": "/metadata/:applicationKey/versionslatest.txt",
+          "responses": {
+            "200": { "type": "string" },
+            "401": { "type": "unit" },
+            "404": { "type": "unit" }
+          }
+        },
+
+        {
+          "method": "GET",
 	  "path": "/:applicationKey",
           "description": "Search all versions of this application. Results are always paginated.",
           "parameters": [
@@ -1366,36 +1391,6 @@
           }
 	}
 
-      ]
-    },
-
-    "application_metadata": {
-      "path": "/:orgKey/metadata/:applicationKey",
-      "operations": [
-        {
-          "method": "GET",
-          "description": "Returns the versions assocoated with the specified application. The latest version is the first result returned.",
-	  "path": "/versions",
-          "parameters": [
-            { "name": "limit", "type": "long", "default": 25, "description": "The number of records to return", "maximum": 100 },
-            { "name": "offset", "type": "long", "default": 0, "description": "Used to paginate. First page of results is 0.", "minimum": 0 }
-          ],
-          "responses": {
-            "200": { "type": "[application_metadata_version]" },
-            "401": { "type": "unit" },
-            "404": { "type": "unit" }
-          }
-        },
-        {
-          "method": "GET",
-          "description": "Returns the latest version number as a string",
-	  "path": "/versions/latest.txt",
-          "responses": {
-            "200": { "type": "string" },
-            "401": { "type": "unit" },
-            "404": { "type": "unit" }
-          }
-        }
       ]
     }
   }
